Fix path of unit test and add code coverage
