+++ .github/workflows/ci-linux.yml
@@ -12,31 +12,53 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v4.1.0
+
       - name: Setup Node.js
         uses: actions/setup-node@v3
         with:
           node-version: 22.11.0
+
       - name: Authenticate with GitHub package registry
         run: echo "//npm.pkg.github.com/:_authToken=${{ secrets.GITHUB_TOKEN }}" > ~/.npmrc
+
       - name: Install dependencies
         run: npm install
+
+      - name: Run unit tests with coverage
+        env:
+          CX_API_KEY: ${{ secrets.API_KEY }}
+          CX_TEST_SCAN_ID: ${{ secrets.TEST_SCAN_ID }}
+        run: |
+          npm run test:coverage
+
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@v3
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }}
+          files: ./coverage/lcov.info
+          flags: unittests
+          name: coverage-report
+          fail_ci_if_error: true
+
       - name: Run tests
         if: runner.os != 'Linux'
         env:
-          CX_API_KEY: ${{ secrets.API_KEY}}
+          CX_API_KEY: ${{ secrets.API_KEY }}
           CX_TEST_SCAN_ID: ${{ secrets.TEST_SCAN_ID }}
         run: |
           npm test
+
       - name: Integration
         if: runner.os == 'Linux'
         env:
-          CX_API_KEY: ${{ secrets.API_KEY}}
+          CX_API_KEY: ${{ secrets.API_KEY }}
           CX_TEST_SCAN_ID: ${{ secrets.TEST_SCAN_ID }}
         run: |
           sudo apt-get install xvfb
           export DISPLAY=:99.0
-          Xvfb -ac :99 -screen 0 1920x1080x16 & 
+          Xvfb -ac :99 -screen 0 1920x1080x16 &
           npm test
+
       - name: Upload test artifacts
         if: ${{ failure() }}
         uses: actions/upload-artifact@c7d193f32edcb7bfad88892161225aeda64e9392 #v4
@@ -45,6 +67,7 @@ jobs:
           path: |
             test-resources/screenshots/*.png
           retention-days: 2
+
       - name: Upload screenshots
         uses: actions/upload-artifact@c7d193f32edcb7bfad88892161225aeda64e9392 #v4
         if: failure() && runner.os == 'macOS'
