+++ .nycrc
@@ -1,0 +1,8 @@
+{
+  "extends": "@istanbuljs/nyc-config-typescript",
+  "all": true,
+  "include": ["src/**/*.ts"],
+  "exclude": ["**/*.test.ts", "mocks/**", "**/*.mock.ts"],
+  "reporter": ["text", "lcov"],
+  "extension": [".ts"]
+}
