+++ package.json
@@ -917,10 +917,12 @@
     "@typescript-eslint/eslint-plugin": "^7.0.1",
     "@typescript-eslint/parser": "^7.2.0",
     "chai": "4.3.1",
+    "codecov": "^3.8.3",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
     "mocha": "10.3.0",
     "mock-require": "^3.0.3",
+    "nyc": "^17.1.0",
     "ts-node": "^10.9.2",
     "typescript": "^5.6.3",
     "vsce": "^2.15.0",
